cmov

Module slice

Trait impls for core slices.

Macros

assert_lengths_eq 🔒

Assert the lengths of the two slices are equal.

assert_size_and_alignment_eq 🔒

Ensure the two provided types have the same size and alignment.

impl_cmov_with_cast 🔒

Implement Cmov and CmovEq traits by casting to a different type that impls the traits.

impl_cmov_with_loop 🔒

Implement Cmov using a simple loop.

impl_cmoveq_with_cast 🔒

Implement CmovEq traits by casting to a different type that impls the traits.

impl_cmoveq_with_loop 🔒

Implement CmovEq using a simple loop.

Constants

WORD_SIZE 🔒

Functions

cast_slice 🔒 ⚠

Performs an unsafe pointer cast from one slice type to the other.

cast_slice_mut 🔒 ⚠

Performs an unsafe pointer cast from one mutable slice type to the other.

cmovne_remainder 🔒

Compare the two remainder slices by loading a Word then performing cmovne.

cmovnz_remainder 🔒

Load the remainder from chunking the slice into a single Word, perform cmovnz, then write the result back out to dst_remainder.

slice_as_chunks 🔒

Rust core [T]::as_chunks vendored because of its 1.88 MSRV. TODO(tarcieri): use upstream function when we bump MSRV

slice_as_chunks_mut 🔒

Rust core [T]::as_chunks_mut vendored because of its 1.88 MSRV. TODO(tarcieri): use upstream function when we bump MSRV

slice_as_chunks_unchecked 🔒 ⚠

Rust core [T]::as_chunks_unchecked vendored because of its 1.88 MSRV. TODO(tarcieri): use upstream function when we bump MSRV

slice_as_chunks_unchecked_mut 🔒 ⚠

Rust core [T]::as_chunks_unchecked_mut vendored because of its 1.88 MSRV. TODO(tarcieri): use upstream function when we bump MSRV

slice_to_word 🔒

Create a Word from the given input slice.

word_to_slice 🔒

Serialize Word as bytes using the same byte ordering as slice_to_word.

Type Aliases

Word 🔒