cmov::slice

Function cmovnz_remainder

fn cmovnz_remainder<T>(
    dst_remainder: &mut [T],
    src_remainder: &[T],
    condition: Condition,
)
where
    T: BitOrAssign + Copy + From<u8> + Shl<usize, Output = T>,
    u64: From<T>,

Load the remainder from chunking the slice into a single Word, perform cmovnz, then write the result back out to dst_remainder.